feat: add VisualizationPanel for enhanced trace event visualization
